+++ packages/react-jss/src/utils/sheets.js
@@ -1,12 +1,11 @@
 // @flow
 import warning from 'tiny-warning'
-import {type StyleSheetFactoryOptions} from 'jss'
+import {getDynamicStyles, type StyleSheetFactoryOptions} from 'jss'
 import type {StyleSheet} from 'jss'
 import type {Context, DynamicRules, Styles} from '../types'
 import {getManager} from './managers'
 import defaultJss from '../jss'
 import {addMeta, getMeta} from './sheetsMeta'
-import getDynamicStyles from './getDynamicStyles'
 
 type Options<Theme> = {
   context: Context,